[update] : useSettings周りを適した形に修正
