+++ src/hooks/index.ts
@@ -1,7 +1,7 @@
 export { default as useResponsive } from './useResponsive';
 export { default as useNotification } from './useNotification';
 export { default as useLocalStorage } from './useLocalStorage';
-export { default as useSetting } from './useSetting';
+export { default as useSettings } from './useSettings';
 export { default as useOverflowDetect } from './useOverflowDetect';
 export { default as useFilteredNotes } from './useFilteredNotes';
 export { default as useSplitView } from './useSplitView';
